+++ classes/trigger_bot.py
@@ -63,6 +63,7 @@ class CS2TriggerBot:
         self.config = config
         self.memory_manager.config = self.config
         self.load_configuration()
+        logger.debug("TriggerBot configuration updated.")
 
     def play_toggle_sound(self, state: bool) -> None:
         """Play a sound when the toggle key is pressed."""
@@ -122,7 +123,6 @@ class CS2TriggerBot:
             return
         # Set the running flag to True and log that the TriggerBot has started
         self.is_running = True
-        logger.info("TriggerBot started.")
 
         # Define local variables for utility functions
         is_game_active = Utility.is_game_active
@@ -164,17 +164,15 @@ class CS2TriggerBot:
                 logger.error(f"Unexpected error in main loop: {e}", exc_info=True)
 
     def stop(self) -> None:
-        """
-        Stops the TriggerBot and cleans up resources.
-        Also stops the keyboard and mouse listeners.
-        """
-        # Set the running flag to False and signal the stop event
+        """Stops the TriggerBot and cleans up resources."""
         self.is_running = False
         self.stop_event.set()
-        # Stop the keyboard and mouse listeners
-        if self.keyboard_listener.running:
-            self.keyboard_listener.stop()
-        if self.mouse_listener.running:
-            self.mouse_listener.stop()
-        # Log that the TriggerBot has stopped
-        logger.info(f"TriggerBot stopped.")
+        time.sleep(0.1)
+        try:
+            if self.keyboard_listener.running:
+                self.keyboard_listener.stop()
+            if self.mouse_listener.running:
+                self.mouse_listener.stop()
+            logger.debug(f"TriggerBot stopped.")
+        except Exception as e:
+            logger.error(f"Error stopping TriggerBot: {e}")