+++ ChangeLog.md
@@ -5,7 +5,10 @@
 - Implemented new GUI tabs for overlay and trigger settings.
 - Integrated new settings into the configuration manager.
 - Improved the general settings tab with new options.
+- Improved configuration updates.
 - Refactored code for better organization and readability.
 - Removed unused image files.
+- Threads optimization.
+- Fixed a bug with the "Start Client" and "Stop Client" buttons.
 
 ![Downloads](https://img.shields.io/github/downloads/Jesewe/VioletWing/v1.2.6/total?style=for-the-badge&logo=github&color=D5006D) ![Platforms](https://img.shields.io/badge/platform-Windows-blue?style=for-the-badge&color=D5006D) ![License](https://img.shields.io/github/license/jesewe/cs2-triggerbot?style=for-the-badge&color=D5006D)